+++ migrations/postgres/20241107130854_outbox_message_version.sql
@@ -1,0 +1,9 @@
+ALTER TABLE outbox_messages ADD version INT;
+
+/* -----------------------------------------------
+Initial one time migration to set message version
+-------------------------------------------------- */
+
+UPDATE outbox_messages set version = 1;
+
+/* ----------------------------------------------- */