+++ src/domain/auth-rebac/services/tests/tests/test_multi_tenant_rebac_dataset_lifecycle_message_consumer.rs
@@ -196,6 +196,7 @@ impl MultiTenantRebacDatasetLifecycleMessageConsumerHarness {
             &self.catalog,
             ConsumerFilter::AllConsumers,
             &content_json,
+            TMessage::version(),
         )
         .await
         .unwrap();