+++ src/domain/task-system/services/tests/tests/test_task_executor_impl.rs
@@ -196,9 +196,10 @@ impl TaskExecutorHarness {
                         })) if task_id == a_task_id
                     )
                 }),
+                eq(1),
             )
             .times(1)
-            .returning(|_, _| Ok(()));
+            .returning(|_, _, _| Ok(()));
 
         mock_outbox
             .expect_post_message_as_json()
@@ -213,9 +214,10 @@ impl TaskExecutorHarness {
                         })) if task_id == a_task_id
                     )
                 }),
+                eq(1),
             )
             .times(1)
-            .returning(|_, _| Ok(()));
+            .returning(|_, _, _| Ok(()));
     }
 
     fn add_run_probe_plan_expectations(