+++ src/infra/core/tests/tests/use_cases/test_append_dataset_metadata_batch_use_case.rs
@@ -221,9 +221,10 @@ impl AppendDatasetMetadataBatchUseCaseHarness {
                         Ok(DatasetLifecycleMessage::DependenciesUpdated(_))
                     )
                 }),
+                eq(1),
             )
             .times(times)
-            .returning(|_, _| Ok(()));
+            .returning(|_, _, _| Ok(()));
     }
 }