+++ src/infra/core/tests/tests/use_cases/test_create_dataset_use_case.rs
@@ -109,9 +109,10 @@ impl CreateUseCaseHarness {
                         Ok(DatasetLifecycleMessage::Created(_))
                     )
                 }),
+                eq(1),
             )
             .times(times)
-            .returning(|_, _| Ok(()));
+            .returning(|_, _, _| Ok(()));
     }
 }