+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Recommendation: for ease of reading, use the following order:
 - Fixed
 -->
 
+## [Unreleased]
+### Changed
+- Add version for `OutboxMessage` structure to prevent startup failures after breaking changes
+
 ## [0.207.2] - 2024-11-15
 ### Fixed
 - E2E: revision of st/mt tests: